Move utility class to util
